\name{print-methods}
\alias{print.JAGS_model}
\alias{print.avg}
\alias{print.gof}

\title{S3 print methods}

\description{
Print S3 objects of class \code{JAGS_model}, \code{avg} and \code{gof}.
}

\usage{
print(x, \dots)
}

\arguments{
  \item{x}{Relevant S3 class object.}
  \item{\dots}{Other arguments to be passed to \code{\link{print}}. Currently ignored.}
}

\author{\email{Brecht.Devleesschauwer@UGent.be}}
